Add arg for worker threads & fix stats init
We now have a configurable number of worker threads although we still
default to the number of detects CPU cores (logical cores a.k.a. threads
really). This makes it possible to just play around with what happens
with lower count or whatever.

The fixed number of max worker threads (32) was rather low. It's bumped
up to 256, now using a single define MAX_WTHREADS so it is consistent
and easy to change. There is also a check that we don't actually try to
run with a higher number.

Alternatively I guess we could malloc this structure but I dunno... this
just felt simple.