Support actonc build on Apple M1 macos/aarch64
This adds support for pointing out the necessary directories for building acton
on Apple M1 silicon, i.e. the aarch64 architecture.

Homebrew uses /opt/homebrew on Apple silicon while it is /usr/local on x86,
thus we specify either based on the architecture (and having detected MacOS).

We are sort of tying ourselves together with Homebrew, but I think that is okay
for now. Not really sure what other mechanisms could have been used to build it
on a x86 mac anyway.