Support building acton programs on Apple M1 / aarch64
When we compile Acton programs we link it with some libraries, most notably
libuuid, which we need to specify the path to. Since Homebrew uses a different
directory (/opt/homebrew) on Apple M1 silicon compared to Intel x86
(/usr/local), we need to adapt accordingly. We opt for using the build time CPP
macros rather than figure this out at run time.

Just like for the build of Acton itself, this means we are now tied to using
Homebrew on Apple M1 but I think that is okay. Not sure what the alternative
would look like anyway. Our longer term goal is to statically link these
libraries in anyway (so they are in libActon.a), so there is no actonc run time
stuff going on.