Skip protobuf generation
I previously added this target as it was required to override the one
above it that is a generic target for all the .c files in the
failure_detector directory. When I added, I implemented an actual target
that generates these .c files from the .proto file.

However, our strategy here is not to generate these files on every
build. Rather, we generate them once and have the output checked into
git. Thus, I've turned this target into a no-op instead.

Anyone wishing to regenerate the files can do so manually and check in
the resulting files in git.