Switch linux test to run on Ubuntu 18.04
We continue building the .deb files on Debian but the standard Linux
test is better run on Ubuntu 18.04, since it is an older platform, so we
cover more ground than running everything on Debian.

The homogeneity we had with debian was mostly because it was simple. Now
that it is working we want to improve coverage across versions etc.