Compile root actor with -no-pie
We don't have a deliberate strategy or approach to position independent
compilation but have relied on defaults of our compilation stack. Acton
is a special case because it includes compilation as part of its run
time, like we compile builtin.o, when building Acton and then later on
use it when running actonc. Compilation flags, such as -pie, need to
align. Since we haven't been using explicit flags, we can get a mismatch.

ISTM we are effectively doing -no-pie, so I'm now adding it. This should
avoid a problem on Linux using Homebrew. It only works on gcc, so we
don't apply this on Macos.