Revamp RTS arg parsing
Rather than rely on getopt and try to wrangle it in various ways to make
it do what we want (we turn off errors and that has weird consequences)
we do our own option parsing. This isn't as bad as it sounds. We don't
have short args, only long ones, so this is fairly basic and while line
count might be up slightly we end up with more straight forward code.

This fixes situations with the previous code that led to segfaults, like
when an option argument is missing. Unlike the earlier code, we now also
correctly handle passing arguments like:
  ./app --rts-wthreads 8
  ./app --rts-wthreads=8

Since we've switched into a single pass parsing rather than two passes
that we previously had (which wasn't really a conscious design, but
happenstance) we end up allocating a slightly larger new_argv. It is the
same size as argv rather than being smaller, since we don't have to pass
all argument. This is negligible.