Make DDB stop more robust
Save the pid to a file so we can kill it specifically rather than solely
relying on pattern matching. I noticed a problem when running on OS
X (have not cross checked on Linux) where the pattern matching failed
due to the command line args not looking as expected. I wonder if this
is because we modify argv when parsing the arguments (I have a
recollection that we do)... pid file is more robust for this case, yet
pattern matching have worked in the past, so I'm keeping it too.