Add DDB test harness
test_db.py is a new script that can orchestrate the setup and testing of
an Acton application together with the distributed database. Multiple
database nodes are started and their status read (from stdout) to ensure
that they agree on their membership before proceeding to run the Acton
application, specifying arguments to use the database backend.

There's an accompanying Acton application, test_db.act, that is written
to be a small and simple test of the database. By calling actor methods
on other actors, we exercise the database, since each call becomes a
message and continuation. We also check that the timer queue works by
invoking one method with after 0:

Resumption of the application is not tested since that is currently
broken. We also do not test running the application with a replication
higher than 1, since that is also broken.

Trying to use a higher replication factor fails but unfortunately this
is not reflected by the application exiting with an error code, rather
we just get some error messages on stderr. The test harness checks for
these, so running with --db-nodes=3 for example will result in a test
failure.

At the very least, this should stop us from introducing new regressions
that affect the use of the database backend from the RTS.

This replaces the earlier hacked up ddb_count test that was driven by a
Makefile, since this new script is better in pretty much all regards.

Changed how tests are invoked in the Makefile as I had problems getting
them all to run when listed as dependencies, so invoking them explicitly
instead. I wonder if this means we have not run some of our tests in the
past?