Serialize env actor during bootstrap
The env actor was never serialized and as a result, for a system that resumes
from the database, we don't have a env actor and any actor method call on env
will fail catastrophically (segfault).

Since the env actor is a sort of constant actor that might not be scheduled for
execution by the main_loop, we might not serialize it. Thus, to fix this, we
simply serialize it just after it has been created.