Test DDB App recovery
This adds a new test case for testing application state recovery from
the database. I initially wanted to write an application that we could
interact with from a TCP connection but due to bugs and not implemented
things, that won't work just yet. Thus I settled on this simple counting
app, which doesn't require any interaction - we can just observe the
output which is easily done from stdout.

We verify application recovery works by starting the app once, letting
it proceed "half way", printing the numbers 0 and 1, then killing it
and upon starting it again, we check that it picks up its count from
where it left of, i.e. that it prints 2 and 3 after which it exits. It
works! Magic!

I quite dislike the counting app because it waits for 1 second between
each count so the whole test takes multiple seconds when it really could
run in a few milliseconds. We can't decrease the wait period since the
after statement currently only supports integer seconds. I'm not sure I
want to decrease it by much anyway since that would make it more timing
sensitive. A better test app would not rely on timing at all, rather we
would interact with the app over like a TCP connection, but that doesn't
work just yet.

I've kept the simple test app too, since it is much faster to run.