Log output during simple DDB App test
While working on the DDB App recovery test I noticed that sometimes the
execution of a DDB app tests could take minutes and I was particularly
surprised that this is true for the simple App! Since we ran that using
subprocess.run(), we would get no output while running and can't really
debug it at all. I have now replaced that with a custom function that
will capture the output of both stdout and stderr, checking for errors
as the program runs. We should thus get a slightly better idea of where
things might be failing. We do get timing information since we output
using log and it is timestamped.

This is not a solution to the real problem but should help debug.

Should probably rewrite the DDB App recovery test to run the program in
this manner too but one step at a time.