Add RTS mon log
This adds a new option of writing the RTS monitor statistics to a log
file. The path to the file is specified with --rts-mon-log-path and the
period at which the statistics is written is specified with
--rts-mon-log-period with an argument in seconds. Default is 30 seconds.
The first entry is written on startup and then with the specified
period. An entry is always written on shutdown so we ensure we capture
all data.

Reworked a bunch of other things as well, in particular the thread
creation for like the 3452 time. Instead of trying to keep track of all
threads in a single array, where threads would be at different positions
depending on if the rts mon functionality was used or not... I've opted
for a simpler, perhaps more verbose, solution... I think this is much
easier to read.