Ensure enough DDB hosts to meet replication factor
We now enforce that the user must specify enough database hosts, using
--rts-ddb-host to reach the replication factor, i.e. if
--rts-ddb-replication is 3, --rts-ddb-host must be specified at least 3
times, like ./app --rts-ddb-host localhost:32000 --rts-ddb-host
localhost:32001 --rts-ddb-host localhost:32002

The idea is that when we connect to a database node, we will communicate
with it and find out about other database nodes that we can connect to,
so that we will reach the required number for the configured replication
factor. With a replication factor of 3, we want to connect to 3 servers!

However, the dissemination of DB servers isn't implemented yet, so the
DB client has to be given all the DB servers, yet we didn't force this,
so setting a higher replication factor than the number of available
servers would lead to an error.

We fixed this with a quick hack, simply assuming that all DB servers are
running on the same machine on sequential ports, so if we specify port
32000 and replication factor 3, we will connect to servers on port
32000, 32001 and 32002. It's an assumption that works while we are doing
development... Naturally, we want to be able to run an actual
distributed system, like on different computers, so we must break free
of this assumption. We now do that by placing the burden on the user
until we can get that gossip protocol implemented!