Do single line log messages
We had some log messages that would print a partial log message on the
start of an operation and print the rest on completion. Since stdout is
line buffered, this meant we had to flush the output and if anything
printed log messages in between the output would look all messed up.

Avoid the complexity by always logging a complete line!