Do not use uuid from util-linux on MacOS
It appears that we might have been compiling with a uuid.h from the
MacOS base / SDK but linked with uuid from util-linu and this might be
the cause of certain segfaults.

Looks like we don't need -luuid on Mac, it works anyway, so trying to
align so we get the same version of headers and library by ditching use
of util-linux libuuid on Mac. We still use it on Linux!