Add script to generate and check env.[ch]
env.[ch] is partially generated based on __builtin__.act but tends to
fall out of sync since it was generated long ago and the definition gets
updated in the compiler which causes the already materialized output in
env.c to drift out of sync.

This new little script is a somewhat naive attempt at helping us
mitigate the problem. It has two modes; --gen and --check

With --gen, it will:
- generate new output from __builtin__.acton
  - extracting the part that pertains to env
- parse the generated output
- parse the existing env.[ch] file
  - note if function is in generated section
    - which means we can overwrite it
- compare generated vs existing
  - for functions, check that return type & arguments align
- use manually defined function if it exists
- otherwise use generated function, possibly overwriting previously
    generated function
- actually do this not just for functions but also other things

--check will do a similar dance, generating new output and comparing it
to the existing content. It will also check structs to ensure that
the fields of manually defined structs align with generated ones up to
the length of the generated one. This means it is possible to manually
define structs with more fields, but the ones that overlap with the
generated one must match! Useful for structs of Actor type where there
is a common set of fields and which has been updated...

The script itself isn't very pretty at all. Parts of it are brittle or
very brittle, yet I believe it will provide value. It has already shown
us mismatching function signatures and that we are lacking a number of
functions in the existing env.[ch] files. My idea here is that we let
the script remain in rather poor form until we've established all use
cases. Once we are happy with those, we can rework sections of the
script in order to improve readability and robustness... or we don't and
just accept this looks ugly and that's fine ;)