Generate parts of env.[ch]
By removing function and struct definitions, the gen_env.py script will
notice they are missing in the file and automatically fill them in the
generated section. I have removed and let the script generate all
functions that I think do not have any form of local modifications.

Some reordering of sections was necessary in order to get ordering of
declarations right.

More notable change is that the generated code does not include __bool__
and __str__ methods. Not sure what's up with that.