Avoid unnecessary wait in comms thread
The comms thread uses select to wait for interesting things to show up,
like messages from servers.

At the very start, when the DB instance in the client API has been
created and thus the comms thread has been started, there are not yet
any servers but the comms thread runs and will select. Since there are
no fds, it will timeout after the timeout of 3 seconds and then take a
second pass in the loop, at which point there likely are some servers
added.

To avoid such waiting on the select loop to time out, we add a wakeup
pipe that can be used to signal the comms thread that something
interesting has happened. We invoke this in the add_server_to_membership
function, so that the comms thread can take another loop and add the new
server to its select readfds set.

This removes 3 seconds from application startup when using the database.