Use netstrings for RTS Monitor interface
Thus far we have used the Luck protocol - that is reading a bit of data
on the socket and just hoping it is enough to either recognize a command
or in the other direction, to get all the data. Last commit on the
actonmon client I had to adjust the recv to get more data so we get it
all.

We still want a simple line protocol, so we are now doing netstrings!
It's just taking the content and prefixing it with the content length
and a colon followed by the content, then ending the whole thing with a
colon.

Our command "WTS" thus becomes "3:WTS," and our JSON response becomes...
well, something very long, but now prefixed with like 20132:{...},

I'm using a simple netstrings library for this. Beyond more robust
reading, we can now also collect multiple commands and execute them all.
We don't really use this, but it could potentially be useful if we
decide to split up the stats data so multiple requests are required in
order to get it all. Having a round trip in between each is sort of bad
since we want all to be captured at the same time, so pipelining
multiple requests and having the RTS answer them in very rapid
succession is probably just as good as what we do today.