Revamp DB interaction in tests
With the new monitor interface on the DB we can use it to determine when
the DB has started and ask about membership. Much better than tailing
stdout and regex parsing its output.

Further, we now compute a random base port which is used when starting
the DB cluster and to feed the right arguments to the Acton RTS.