Add TCP based app recovery test
This adds an app recovery test, where we want to check that an
application has correctly recovered its state from the database upon
restart. Unlike the existing app recovery test, this one is based on a
TCP server, which makes it much faster, more robust and avoids the
timing issues that we have with the current app recovery test.

The count app outputs a number every second. We cannot lower the sleep
in between numbers, partly because after X doesn't take integer seconds
as arguments but also because we depend on killing the application
precisely when it has counted to 3. We then expect it to continue
counting from there. If the app is too fast, we might get racy
conditions and end up killing the app too late.

Since we now interact over TCP, a query completes in "no time at all"
and the whole test that runs a few calls rounds down to "no time at all"
as well ;) So beyond the improvement in robustness by talking over TCP,
it is sooooo much faster.

Unfortunately, the RTS doesn't properly support recovery of TCP sockets,
so that particular part of the test is currently disabled. We should
have a socket error handler that we install in the env.listen() call,
like we should do:

    env.listen(port, on_connect, on_error)

If there's a problem establishing the listening socket or there is ever
a problem later on with the socket (such as when doing recovery from
DB), we can call the on_error callback and let the application handle
it. See #440 for details.