Separate env.listen error callback
env.listen(port, cb) has previously taken a single callback cb that is
invoked both when there are incoming connections to the listening socket
as well as for errors, in which case the callback was called with a None
argument. This is confusing, bad, bla bla. Better to have a separate
callback, which we are now introducing, so the new signature is:

    env.listen(port, cb_on_connect, cb_on_error)

An error when trying to establish the listening socket will now be
reported by calling the callback function provided via cb_on_error.

There's a new test case to cover precisely this test case.