Add actor resume method
The __resume__ method is run immediately after an actor has been
deserialized from the database upon resumption of the system and allows
us to add extra hooks when necessary. For example, for socket related
actors, we want to invoke the error callback so that the user
application code can recreate the socket.