Call error handler on resume of ListenSocket
When a ListenSocket actor is resumed from the database, its error
handler is now called so that the Acton application is notified that the
ListenSocket is no longer working.

We unconditionally call the error handler because there is no way that
the ListenSocket survives a resume. We serialize it many times to the
database in normal operations but we only deserialize and resume it when
we start up fresh (or in the future when we migrate an actor). Either
way, the listen socket has a local fd in the kernel and that will not
survive either circumstance so it assuming it broken and reporting via
the on_error callback is the only sane recourse. The Acton application
should have an error handler that sets up the listening socket anew.

We really should not need to create the cb_err attribute, the actor
argument cb_on_error should be saved and available, however the compiler
currently prunes it away. It runs an analysis to see if actor arguments
are used by methods in the actor, in which case they are persisted.
Since this is a abstract actor with stub methods, the analysis comes up
with the wrong result that the arguments are not needed. We will try to
address this separately, see #524.