Improve DB test retry
Besides interacting over TCP we check the process itself to detect
segfaults or other early and unexpected exits. The retry sleep logic is
improved with a backoff. This is necessary to detect segfaults on my
MacBook Pro. Writing the core file, which takes 3.5GB, takes several
seconds and meanwhile the process is still "live" (returncode = None).
By waiting it out, we can finally see returncode -11, i.e. a segfault
and report that in the test.

Also compile test app with --dev to get debug symbols so we can get
backtrace after segfaults.

By reducing the port range we're able to reduce the amount of errors
typically seen when running tests. I'm not sure why this is. On my
computer I'm able to reduce it roughly 10x, from something like 1 in 10
to roughly 1 in 100. I think we are colliding with client sockets using
higher ports?