Workaround compiler actor init bug
All actor attributes might not be initialized / zeroized when the actor
is serialized to the database for the first time, which leads to
segfaults. The fix should clearly be applied to the compiler (see #525)
but meanwhile we can get our tests passing by ensuring that all actor
attributes are set before the actor is serialized.