Remove old RTS DDB recovery test
The first test case for verifying the resumption of an application
using the RTS in distributed database mode operation was based on
interacting via stdout. It was hacky and brittle. We now have a much
more robust test application which we communicate to using TCP. Thus, we
can remove the old test and hopefully avoid some intermittent failures.