Fix DB test retry sleep
Do a short sleep between retries. Being too aggressive is not good so we
do a bit of a backoff so that things may have some time to start up /
settle before the next attempt.

Also add a case for when we get no response. I haven't been able to spot
this during my local testing but it doesn't hurt having here in case it
happens in CI etc.

Seeing quite a few segfaults due to various reasons. Will need to debug
that separately... but that's not caused by the changes on this branch,
rather they've always been there, it's just that we can run our test
fast enough and thus repeat the test enough that we notice more
intermittent failures now.