Always pass Connection to on_connect callback
Previously env.listen only took a single callback which was used to
signal both new client connections as well as errors in establishing the
listening socket. The argument would either be a Connection or None to
indicate an error. Now that we have separated errors into its own
callback we can always pass a Connection to the on_connect callback
handler. This makes it easier and cleaner to write application code.