Pass connection argument to on_receive
We install a receive handler on a Connection actor by calling the
on_receive() method and when a message is received by the Connection it
will execute this callback. The only argument passed used to be the
message content itself, which could make it somewhat difficult to
understand on which Connection the message was received. We now pass
both the connection and the message so that the handler function can
more easily understand which client Connection is being dealt with.