Test TCP client
I've added a new test case that uses a TCP client. It's really meant to
test resume support of the Connection actor (and that it calls the error
handler, which in turn should do a reconnect), however that stuff is not
implemented yet, so I'm sort of abusing this test to test basic TCP
client behavior for now. Once TCP client connection resume is working,
we can just uncomment the relevant lines.