Add simple TCP server stress test case
In order to exercise our TCP machinery for a bit, we now add a new test
that starts the ddb_test_server and then connects 2000 times to it. The
client connections are run sequentially, so it's not really stressing
stuff but at least we can see how we must clean up fds, otherwise the
program just stops at the fd limit (1024 per default).