Use log.c lib in ActonDB
Rather than printf() log message, we now use the same log.c library that
is used in RTS, though with a slightly different set of modifications.

Not all printf() statements have been replaced but most have. I've
mapped most log lines to log_debug() but have lowered the level of some
messages to log_info(). Rather than scrub all this in one go, I figured
we could over time adjust the levels as we see fit.

It might seem weird to have two log libraries and in particular with
different set of modifications but that was quite deliberate, since the
use in the DB and RTS are quite different. For example, RTS is threaded
while the DB is not (although it could become). The RTS log.c prints the
worker thread id, so it's possible to track what log lines belong to
what worker thread. The DB doesn't have worker threads, so such a change
would otherwise have to be made conditional. Keeping them separate is
thus, for the time being, simpler. There might be more changes that we
want to apply. If we see the two log libs being rather stable over time
and can find a way to combine, that would be fine too.