Build stdlib modules using actonc
Thus far, all stdlib modules have been implemented in C and so the
compilation of them has been through manually defined Makefile targets.

We now support compiling modules written in Acton!

actonc doesn't play very well with Make since given a source file as
input it produces multiple output files. When we have make targets for
all the output files but with the same build rule (actonc $<), we'll
potentially have multiple actonc invocations concurrently. We also need
to compile the stdlib with and without the --dev flag, which increases
concurrency further. Multiple invokations will step on each other and
crash with compiler errors.

The longer term plan is for actonc to support concurrent builds and
building an Acton project should be a simple matter of "actonc build".
We're not quite there yet, so in the meantime we have to invoke actonc
multiple times but ensure we only invoke actonc once at a time, thus the
stdlib_project target which looks a little funny and is defined as a
phony.

Despite being phony, we manage to get the dependencies right and not
pollute our non-phony targets (since stdlib_project is never used as a
prerequisite).