Move C parts of time module to _time
The time module, which is implemented in C, has been renamed to _time
and a new time module with wrapper functions for all the existing
functions in _time has been created. Thus the interface remains the same
but the time module is now implemented in Acton, making it possible to
add more things implemented in Acton! This is preparatory for planned
improvements to the time module.