Add module dependencies
actonc is able to recompile imported modules, like if we compile
numpy.act, which imports math, actonc will also compile math, if it was
not already compiled.

When we don't have our dependencies right, we'll get races because Make
might invoke compilation of numpy and math at the same time and since
compiling numpy implies compiling math, we get a collision.

Long term we still want actonc to deal with all of this. Right now, we
compute which .ty and .h files are stub compiled and add those as
dependencies to the stdlib_project target, so all C modules will
effectively be compiled first, only then do we compile the acton native
modules in stdlib.