Infer stub mode compilation on presence of .c file
We now automatically infer stub mode compilation if there is a .c file
with the same filename (ignoring the extension) as the act file. For
example, if we are compiling _time.act and _time.c exists, we will
assume stub mode compilation.

Chased import files are stub checked separately so that the main target
can be normally compiled while one of its dependencies can be stub
compiled.

--stub has been removed in a number of places in the Makefile since we
can now detect it. Notably __builtin__.act is the only file that still
has it, since we compile the C files differently and there is no
__builtin__.c