Hardcode stub compilation for __builtin__.act
Since we're already hard coding stuff for __builtin__.act it doesn't
hurt hard-coding some more, in this case stub mode compilation.

We no longer need to specify --stub anywhere in our Makefile, which
means that we could potentially remove the --stub argument completely
but as I'm uncertain on whether there might be other weird edge cases
where it could be useful, I'm leaving it in for now.

This is also in preparation for the actonc build command where we won't
be able to specify --stub specifically for __builtin__.act since we are
just running build for the entire stdlib project.