Correctly implement stub mode
The environment (env) record has a stub field which controls whether
stub compilation is used or not.

We used to initialize this value in initEnv and keep that value in the
env throughout the current actonc invokation, affecting all compiled
files. It was only the initial file that would really affect whether
stub compilation was used or not and later chased imports would use that
same initial value. Since actonc has only supported compiling single
files, that was fine but with the introduction of acton projects we want
to compile many files of which some need stub mode compilation.

Now we instead change the stub mode in the local environment used for
the compilation in runRestPasses, effectively setting stub mode per
file. It means it's now irrelevant for initEnv, so the arg is removed
from there.

Also improved output a bit.