Move __builtin__.act
It's a builtin and not really part of the stdlib! It's a challenge
getting the build of the stdlib acton project to work when all modules
have an implicit dependency on builtins and thus, it needs to be
compiled first.

Better to move the __builtin__.act - all of the C code is in builtin/,
so might as well keep the .act file here too!

An acton project directory structure is used to get the output where we
want it!